+++ aa_forum/views/personal_messages.py
@@ -18,17 +18,15 @@ from django.utils.translation import gettext as _
 # Alliance Auth
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.logging import LoggerAddTag
-
 # AA Forum
 from aa_forum import __title__
 from aa_forum.forms import NewPersonalMessageForm, ReplyPersonalMessageForm
 from aa_forum.helper.pagination import get_paginated_page_object
 from aa_forum.helper.user import get_main_character_from_user
 from aa_forum.models import PersonalMessage, Setting
+from aa_forum.providers import AppLogger
 
-logger = LoggerAddTag(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
 
 
 @login_required